Refactor: Move the serde tests to the appropriate module.
