+++ src/node/serde.rs
@@ -145,3 +145,50 @@ where
         Ok(Nodes(nodes))
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_node_serialize() {
+        let node = Node::new(1, Some(2));
+        let serialized = serde_json::to_string(&node).unwrap();
+        #[cfg(not(feature = "compact_serde"))]
+        assert_eq!(
+            serialized,
+            r#"{"node_id":1,"value":2,"children":[],"parent":null}"#
+        );
+        #[cfg(feature = "compact_serde")]
+        assert_eq!(serialized, r#"{"node_id":1,"value":2,"parent":null}"#);
+    }
+
+    #[test]
+    fn test_node_deserialize() {
+        let node = Node::new(1, Some(2));
+        let serialized = serde_json::to_string(&node).unwrap();
+        let deserialized: Node<i32, i32> = serde_json::from_str(&serialized).unwrap();
+        assert_eq!(node, deserialized);
+    }
+
+    #[test]
+    fn test_nodes_serialize() {
+        let nodes = Nodes::new(vec![Node::new(1, Some(2))]);
+        let serialized = serde_json::to_string(&nodes).unwrap();
+        #[cfg(not(feature = "compact_serde"))]
+        assert_eq!(
+            serialized,
+            r#"[{"node_id":1,"value":2,"children":[],"parent":null}]"#
+        );
+        #[cfg(feature = "compact_serde")]
+        assert_eq!(serialized, r#"[{"node_id":1,"value":2,"parent":null}]"#);
+    }
+
+    #[test]
+    fn test_nodes_deserialize() {
+        let nodes = Nodes::new(vec![Node::new(1, Some(2))]);
+        let serialized = serde_json::to_string(&nodes).unwrap();
+        let deserialized: Nodes<i32, i32> = serde_json::from_str(&serialized).unwrap();
+        assert_eq!(nodes, deserialized);
+    }
+}
